+++ nodeconductor/cost_tracking/managers.py
@@ -1,8 +1,8 @@
 import datetime
 
 from django.contrib.contenttypes.models import ContentType
-from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
-from django.db import models as django_models, IntegrityError
+from django.core.exceptions import ObjectDoesNotExist
+from django.db import models as django_models
 from django.db.models import Q
 from django.utils import timezone
 
@@ -49,24 +49,11 @@ class PriceEstimateManager(GenericKeyMixin, UserFilterMixin, django_models.Manag
 
     def get_current(self, scope):
         now = timezone.now()
-        try:
-            return self.get(scope=scope, year=now.year, month=now.month)
-        except MultipleObjectsReturned:
-            return self.get(scope=scope, year=now.year, month=now.month, is_manually_input=True)
+        return self.get(scope=scope, year=now.year, month=now.month)
 
-    def create_or_update(self, scope, **defaults):
-        """
-        We don't use built-in update_or_create method because
-        we need to deal with manually input and auto-generated price estimates.
-
-        1. If price estimate for current date does not exist yet, we generate it.
-        2. If both auto-generated and manually input price estimates exist, we should update both.
-        """
-        today = timezone.now()
-        try:
-            self.create(scope=scope, year=today.year, month=today.month, is_manually_input=False, **defaults)
-        except IntegrityError:
-            self.filter(scope=scope, year=today.year, month=today.month).update(**defaults)
+    def get_or_create_current(self, scope):
+        now = timezone.now()
+        return self.get_or_create(scope=scope, month=now.month, year=now.year)
 
 
 class ConsumptionDetailsQuerySet(django_models.QuerySet):
@@ -98,7 +85,3 @@ class PriceListItemManager(GenericKeyMixin, UserFilterMixin, django_models.Manag
     def get_available_models(self):
         """ Return list of models that are acceptable """
         return Service.get_all_models()
-
-
-class ResourcePriceItemManager(GenericKeyMixin, django_models.Manager):
-    pass