+++ nodeconductor/cost_tracking/tests/factories.py
@@ -2,9 +2,8 @@ import factory
 
 from django.core.urlresolvers import reverse
 from django.contrib.contenttypes.models import ContentType
-from django.utils import timezone
 
-from nodeconductor.cost_tracking import models, CostTrackingStrategy
+from nodeconductor.cost_tracking import models, CostTrackingStrategy, ConsumableItem
 from nodeconductor.structure.tests import models as test_models
 from nodeconductor.structure.tests import factories as structure_factories
 
@@ -91,11 +90,40 @@ class PriceListItemFactory(AbstractPriceListItemFactory):
 class TestNewInstanceCostTrackingStrategy(CostTrackingStrategy):
     resource_class = test_models.TestNewInstance
 
+    class Types(object):
+        STORAGE = 'storage'
+        RAM = 'ram'
+        CORES = 'cores'
+        QUOTAS = 'quotas'
+        FLAVOR = 'flavor'
+
     @classmethod
-    def get_consumables(cls, resource):
-        return {
-            'disk': resource.disk if resource.state != test_models.TestNewInstance.States.ERRED else 0,
-            'ram': resource.ram if resource.runtime_state == 'online' else 0,
-            'cores': resource.cores if resource.runtime_state == 'online' else 0,
-            'test_quota': resource.quotas.get(name=test_models.TestNewInstance.Quotas.test_quota).usage,
+    def get_configuration(cls, resource):
+        States = test_models.TestNewInstance.States
+        if resource.state == States.ERRED:
+            return {}
+        resource_quota_usage = resource.quotas.get(name=test_models.TestNewInstance.Quotas.test_quota).usage
+        consumables = {
+            ConsumableItem(item_type=cls.Types.STORAGE, key='1 MB'): resource.disk,
+            ConsumableItem(item_type=cls.Types.QUOTAS, key='test_quota'): resource_quota_usage,
         }
+        if resource.runtime_state == 'online':
+            consumables.update({
+                ConsumableItem(item_type=cls.Types.RAM, key='1 MB'): resource.ram,
+                ConsumableItem(item_type=cls.Types.CORES, key='1 core'): resource.cores,
+            })
+        if resource.flavor_name:
+            consumables[ConsumableItem(item_type=cls.Types.FLAVOR, key=resource.flavor_name)] = 1
+        return consumables
+
+    @classmethod
+    def get_configuration_default_prices(cls):
+        return [
+            ConsumableItem(cls.Types.STORAGE, "1 MB", units='MB', name='Storage'),
+            ConsumableItem(cls.Types.RAM, "1 MB", units='MB', name='RAM', default_price=1),
+            ConsumableItem(cls.Types.CORES, "1 core", name='Cores'),
+            ConsumableItem(cls.Types.QUOTAS, "test_quota", name='Test quota'),
+            ConsumableItem(cls.Types.FLAVOR, "small", name='Small flavor'),
+            ConsumableItem(cls.Types.FLAVOR, "medium", name='Medium flavor'),
+            ConsumableItem(cls.Types.FLAVOR, "large", name='Large flavor'),
+        ]