+++ CHANGELOG.rst
@@ -3,6 +3,9 @@ CHANGELOG
 
 Release NEXT
 ------------
+- Rewrite price estimate calculation logic.
+- Do not allow to edit price estimates manually.
+- Store resources consumption details.
 - Fix scoped service settings descendant resource unlinking.
 - Enable HTTP client debugging.